small refactoring: moved the motor diagnostic parser to its own set of files - will be cleaner when we work on issues #11 and #14
