+++ sr_self_test/include/sr_self_test/diagnostic_parser.hpp
@@ -247,37 +247,6 @@ namespace shadow_robot
     };
   };
 
-  class MotorDiagnostics
-    : public MinMaxDiagnostics
-  {
-  public:
-    MotorDiagnostics(std::string name)
-      : MinMaxDiagnostics(name)
-    {
-      values_.reset(new DiagMap() );
-      std::pair<std::vector<DiagValues>, std::vector<DiagValues> > voltage;
-      voltage.second.resize(2);
-      voltage.second[0] = 23.5; //min
-      voltage.second[1] = 24.5; //max
-      values_->insert( std::pair<std::string, std::pair<std::vector<DiagValues>, std::vector<DiagValues> > >("Measured Voltage", voltage) );
-
-      std::pair<std::vector<DiagValues>, std::vector<DiagValues> > temperature;
-      temperature.second.resize(2);
-      temperature.second[0] = 20.0; //min
-      temperature.second[1] = 50.0; //max
-      values_->insert( std::pair<std::string, std::pair<std::vector<DiagValues>, std::vector<DiagValues> > >("Temperature", temperature) );
-    }
-
-    ~MotorDiagnostics()
-    {};
-
-    virtual std::auto_ptr<BaseDiagnostics> shallow_clone(std::string name)
-    {
-      std::auto_ptr<BaseDiagnostics> tmp( new MotorDiagnostics(name) );
-      return tmp;
-    };
-  };
-
   class IsOKDiagnostics
     : public BaseDiagnostics
   {
