+++ sr_self_test/include/sr_self_test/motor_test.hpp
@@ -1,0 +1,53 @@
+/**
+ * @file   motor_test.hpp
+ * @author Ugo Cupcic <ugo@shadowrobot.com>
+ * @date   Mon Apr 22 05:43:10 2013
+ *
+ *
+ * Copyright 2011 Shadow Robot Company Ltd.
+ *
+ * This program is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the Free
+ * Software Foundation, either version 2 of the License, or (at your option)
+ * any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ *
+ * @brief Run different tests on the motors (parses diagnostics, apply PWM offset
+ *        and measure current, test strain gauges, etc.. )
+ *
+ *
+ */
+
+#ifndef _MOTOR_TEST_H_
+#define _MOTOR_TEST_H_
+
+#include "sr_self_test/diagnostic_parser.hpp"
+
+namespace shadow_robot
+{
+  class MotorDiagnostics
+    : public MinMaxDiagnostics
+  {
+  public:
+    MotorDiagnostics(std::string name);
+    ~MotorDiagnostics();
+
+    virtual std::auto_ptr<BaseDiagnostics> shallow_clone(std::string name);
+  };
+}
+
+  /* For the emacs weenies in the crowd.
+     Local Variables:
+     c-basic-offset: 2
+     End:
+  */
+
+#endif /* _MOTOR_TEST_H_ */
